Move TextUtilities into
The static  methods defined in `TextUtilities` were only ever used in
`TextGeneratorComponent` so moved them there.

Created a `Constants` file too, to hold constants. We can add more as
needed.
