+++ src/explorer/Common/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace Explorer.Common
+{
+    public static class Constants
+    {
+        public const string EmailAddressChars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ-_.";
+    }
+}
